heap: match size for length field
The front-ing structure and the implementation had drifted in the size of
the length parameter.

+++ contrib/egg/egg-heap.h
@@ -33,7 +33,7 @@ typedef struct _EggHeap EggHeap;
 struct _EggHeap
 {
   gchar *data;
-  guint  len;
+  gsize  len;
 };
 
 GType      egg_heap_get_type      (void);
